requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
pandas==2.0.3
PyPDF2==3.0.1
anthropic==0.52.1
python-dotenv==1.0.0
Werkzeug==2.3.7
